requirements.txt
django
mysql.connector
mysql-connector-python
requests
